Separate out manifest nodes into individual files
